refactor(handlers): 统一 Handler 文件命名为 kebab-case 格式
- 为什么改：项目中 Handler 文件命名不一致，部分使用 PascalCase（如 ConfigApiHandler.ts），部分使用 kebab-case（如
static-file.handler.ts），为统一命名规范并提升代码可维护性
- 改了什么：
  - 重命名 7 个 Handler 源文件：ConfigApiHandler、HeartbeatHandler、RealtimeNotificationHandler、ServiceApiHandler、
StatusApiHandler、UpdateApiHandler、VersionApiHandler → 对应的 kebab-case 格式（如 config.handler.ts）
  - 重命名 6 个测试文件：保持与源文件一致的命名规范
  - 更新 handlers/index.ts：修改所有导出语句的路径引用
  - 更新 WebServer.test.ts：修改 6 处 vi.mock 的路径
  - 更新 basic.test.ts：修改动态导入路径
  - 更新各个测试文件的导入语句（7 个文件）
- 影响范围：仅影响文件命名和导入路径，类名保持不变，对外 API 无影响，无兼容性问题
- 验证方式：
  - 类型检查通过：pnpm check:type ✓
  - Lint 检查通过：pnpm lint ✓
  - 所有测试通过：567 tests passed ✓
