+++ apps/backend/handlers/heartbeat.handler.ts
