+++ apps/backend/handlers/index.ts
@@ -1,14 +1,14 @@
-export * from "./ConfigApiHandler.js";
+export * from "./config.handler.js";
 export { CozeHandler } from "./coze.handler.js";
-export * from "./HeartbeatHandler.js";
+export * from "./heartbeat.handler.js";
 export { EndpointHandler } from "./endpoint.handler.js";
 export * from "./mcp.handler.js";
 export { MCPHandler } from "./mcp-manage.handler.js";
-export * from "./RealtimeNotificationHandler.js";
-export * from "./ServiceApiHandler.js";
+export * from "./realtime-notification.handler.js";
+export * from "./service.handler.js";
 export { StaticFileHandler } from "./static-file.handler.js";
-export * from "./StatusApiHandler.js";
+export * from "./status.handler.js";
 export * from "./mcp-tool.handler.js";
 export * from "./mcp-tool-log.handler.js";
-export * from "./UpdateApiHandler.js";
-export * from "./VersionApiHandler.js";
+export * from "./update.handler.js";
+export * from "./version.handler.js";
