+++ apps/backend/handlers/version.handler.ts
